feat : sendChat customHook
