+++ src/hooks/useScrollDown.tsx
@@ -1,13 +1,21 @@
-import { useEffect, useRef } from 'react';
+import { throttle } from 'lodash';
+import { useEffect, useRef, useState } from 'react';
 
 const useScrollDown = (dep: any) => {
   const bodyRef = useRef<HTMLDivElement>(null);
+
+  const onScroll = throttle(() => {}, 100);
 
   const goToBottom = () => {
     if (bodyRef.current) {
       bodyRef.current.scrollTo(0, bodyRef.current.scrollHeight);
     }
   };
+
+  useEffect(() => {
+    window.addEventListener('scroll', onScroll);
+    return () => window.removeEventListener('scroll', onScroll);
+  }, []);
 
   useEffect(() => {
     goToBottom();
