+++ src/states/mocks/friendsData.json
@@ -3,31 +3,31 @@
     {
       "userId": 10,
       "userName": "박명수",
-      "profileImage": "https://i.ibb.co/SK2x480/2022-05-05-10-47-58.png",
+      "profileImage": "https://i.ibb.co/MVmZVKM/park.webp",
       "statusMessage": "북쪽에 계신 메리메리"
     },
     {
       "userId": 11,
       "userName": "데프콘",
-      "profileImage": "https://i.ibb.co/Jdmk2VM/2022-05-10-10-04-21.png",
+      "profileImage": "https://i.ibb.co/DLFV9yz/con.webp",
       "statusMessage": ""
     },
     {
       "userId": 12,
       "userName": "길성준",
-      "profileImage": "https://i.ibb.co/Wnt3NFK/2022-05-11-8-04-28-1.png",
+      "profileImage": "https://i.ibb.co/jLk9XQy/gill.webp",
       "statusMessage": "나 이런게 무서워하네"
     },
     {
       "userId": 13,
       "userName": "정준하",
-      "profileImage": "https://i.ibb.co/0CT39hS/2022-05-11-8-19-41.png",
+      "profileImage": "https://i.ibb.co/N2YdRvZ/jung.webp",
       "statusMessage": "웃지마"
     },
     {
       "userId": 14,
       "userName": "유재석",
-      "profileImage": "https://i.ibb.co/RgBCJtd/2022-05-12-2-55-06.png",
+      "profileImage": "https://i.ibb.co/jkppnqL/you.webp",
       "statusMessage": ""
     }
   ]
